added :PATCH /me :)

+++ src/Controller/ProfilController.php
@@ -13,6 +13,7 @@ use App\Http\JsonBodyResponse;
 use App\Model\Table\UsersTable;
 use App\Network\Exception\UnprocessedEntityException;
 use App\Validator\DataValidator;
+use Cake\I18n\FrozenDate;
 use Eswipe\Model\UserInfo;
 
 /**
@@ -48,18 +49,55 @@ class ProfilController extends ApiV1Controller
         if (!is_null($message)) {
             throw new UnprocessedEntityException($message);
         }
-        $userPatch = $this->request->getData();
+        $userPatch = array_filter($this->request->getData(), function ($key) {
+            return $key !== null;
+        });
 
         $user = $this->Users->get($user_id, ['contain' => ['LookingFor']]);
-        $this->Users->LookingFor->unlink($user, $user->looking_for); // suppresion des anciennes liaisons
+        if (isset($userPatch['looking_for'])) {
+            $this->Users->LookingFor->unlink($user, $user->looking_for); // suppresion des anciennes liaisons
 
-        debug($userPatch['looking_for']);
-        $genders = $this->Users->LookingFor
-            ->find('all')->where(['name IN' => $userPatch['looking_for']])->all();
+            $genders = $this->Users->LookingFor
+                ->find('all')->where(['name IN' => $userPatch['looking_for']])->toArray();
 
-        debug($genders);
-        $this->Users->LookingFor->link($user, $genders);
-        debug($user);
+            $this->Users->LookingFor->link($user, $genders);// reassociation de la liaison :)
+        }
+
+
+        if (isset($userPatch['first_name'])) {
+            $user->firstname = $userPatch['first_name'];
+        }
+
+        if (isset($userPatch['last_name'])) {
+            $user->lastname = $userPatch['last_name'];
+        }
+
+        if (isset($userPatch['date_of_birth'])) {
+            $user->date_of_birth = FrozenDate::parseDate($userPatch['date_of_birth']);
+        }
+
+        if (isset($userPatch['description'])) {
+            $user->description = $userPatch['description'];
+        }
+
+        if (isset($userPatch['gender'])) {
+            $user->gender = $this->Users->Genders->findByName($userPatch['gender'])->first();;
+        }
+
+        if (isset($userPatch['looking_for_age_min'])) {
+            $user->min_age = $userPatch['looking_for_age_min'];
+        }
+
+        if (isset($userPatch['looking_for_age_max'])) {
+            $user->max_age = $userPatch['looking_for_age_max'];
+        }
+
+        if (isset($userPatch['is_visible'])) {
+            $user->is_visible = $userPatch['is_visible'];
+        }
+
+        $this->Users->save($user);
+        return $this->response->withStatus(204);
 
     }
